refactor display components
