+++ src/window_main/components/collection/filters.tsx
@@ -4,7 +4,9 @@ import React from "react";
 import { ColumnInstance, Row } from "react-table";
 import { CARD_RARITIES, MANA } from "../../../shared/constants";
 import db from "../../../shared/database";
-import { RaritySymbol, SetSymbol, TypeSymbol } from "../misc/display";
+import { RaritySymbol } from "../misc/RaritySymbol";
+import { SetSymbol } from "../misc/SetSymbol";
+import { TypeSymbol } from "../misc/TypeSymbol";
 import { BinaryColumnFilter, BinaryFilterValue } from "../tables/filters";
 import { useMultiSelectFilter } from "../tables/hooks";
 import { MultiSelectFilterProps } from "../tables/types";
