+++ src/window_main/components/decks/DecksArtViewRow.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { Cell } from "react-table";
 import { CSSTransition } from "react-transition-group";
 import { getCardArtCrop } from "../../../shared/util";
-import { ArtTile, MetricText } from "../misc/display";
+import { ArtTile } from "../misc/ArtTile";
+import { MetricText } from "../misc/MetricText";
 import { DecksData, DecksTableRowProps } from "./types";
 
 function DeckArt({ url }: { url: string }): JSX.Element {
