+++ src/window_main/components/matches/MatchesTableControls.tsx
@@ -4,7 +4,8 @@ import { MATCHES_TABLE_MODES } from "../../../shared/constants";
 import ReactSelect from "../../../shared/ReactSelect";
 import { getReadableEvent } from "../../../shared/util";
 import DateFilter from "../../DateFilter";
-import { MediumTextButton, SmallTextButton } from "../misc/display";
+import { MediumTextButton } from "../misc/MediumTextButton";
+import { SmallTextButton } from "../misc/SmallTextButton";
 import ColumnToggles from "../tables/ColumnToggles";
 import { GlobalFilter } from "../tables/filters";
 import PagingControls from "../tables/PagingControls";
