+++ src/window_main/components/misc/CalendarSymbol.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { SymbolBase } from "./SymbolBase";
+
+export const CalendarSymbol = styled(SymbolBase).attrs(props => ({
+  className: `icon_2 ${props.className ?? ""}`
+}))``;
