+++ src/window_main/components/misc/CheckboxContainer.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const CheckboxContainer = styled.label.attrs(props => ({
+  className: (props.className ?? "") + " check_container hover_label"
+}))`
+  display: inline-flex;
+`;
