+++ src/window_main/components/misc/LabelText.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const LabelText = styled.div`
+  display: inline-block;
+  text-align: left;
+  white-space: nowrap;
+  color: var(--color-light);
+`;
