+++ src/window_main/components/misc/MediumTextButton.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { SmallTextButton } from "./SmallTextButton";
+
+export const MediumTextButton = styled(SmallTextButton)`
+  width: 180px;
+`;
