+++ src/window_main/components/misc/MetricText.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const MetricText = styled.div`
+  display: inline-block;
+  line-height: 32px;
+  font-family: var(--sub-font-name);
+  color: var(--color-light);
+  font-weight: 300;
+`;
