+++ src/window_main/components/misc/SmallTextButton.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { MetricText } from "./MetricText";
+
+export const SmallTextButton = styled(MetricText).attrs(props => ({
+  className: (props.className ?? "") + " button_simple"
+}))`
+  margin: 0 4px 5px 4px;
+  width: 90px;
+`;
