+++ src/window_main/components/misc/SymbolBase.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const SymbolBase = styled.div`
+  line-height: initial;
+  height: 20px;
+  width: 20px;
+  display: inline-block;
+  background-size: contain;
+  background-position: center;
+  margin: auto 2px;
+  vertical-align: middle;
+`;
