+++ src/window_main/components/misc/TicketSymbol.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { SymbolBase } from "./SymbolBase";
+
+export const TicketSymbol = styled(SymbolBase).attrs(props => ({
+  className: `economy_ticket ${props.className ?? ""}`
+}))``;
